Fix for colors counted more than 256 times in one lump

+++ src/i_colors.c
@@ -182,7 +182,7 @@ int FindDominantBrightColor(patch_t *patch, byte *palette)
     const int   w = SHORT(patch->width);
     int         dominantcolor = 0;
     int         dominantcolorcount = 1;
-    byte        colorcount[256] = { 0 };
+    int         colorcount[256] = { 0 };
 
     for (int col = 0; col < w; col++)
     {
@@ -221,7 +221,7 @@ int FindDominantEdgeColor(patch_t *patch)
 {
     int         dominantcolor = 0;
     int         dominantcolorcount = 1;
-    byte        colorcount[256] = { 0 };
+    int         colorcount[256] = { 0 };
     column_t    *column = (column_t *)((byte *)patch + LONG(patch->columnofs[0]));
 
     while (column->topdelta != 0xFF)
